backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
celery
redis
cryptography
python-jose[cryptography]
passlib[bcrypt]
beautifulsoup4
firebase-admin
